Refactor Code Base
- [+] feat(NewStuffLLMs.ts): add function to retrieve configuration for new stuff based on model
- [+] feat(NewStuffLLMs.ts): add function to get corresponding model name for instruct versions
- [+] fix(openai.ts): remove duplicate implementation of getNewStuff and getModelForInstructVersion functions
- [+] refactor(openai.ts): update references to getNewStuff and getModelForInstructVersion functions
