+++ app/client/platforms/openai.ts
@@ -15,6 +15,10 @@ import {
   fetchEventSource,
 } from "@fortaine/fetch-event-source";
 import { sendModerationRequest } from './textmoderation';
+import { 
+  getNewStuff,
+  getModelForInstructVersion,
+} from './NewStuffLLMs';
 import { prettyObject } from "@/app/utils/format";
 import { getClientConfig } from "@/app/config/client";
 import { getProviderFromState } from "@/app/utils";
@@ -66,59 +70,6 @@ export class ChatGPTApi implements LLMApi {
 
   extractMessage(res: any) {
     return res.choices?.at(0)?.message?.content ?? "";
-  }
-
-  /**
-   * Retrieves information about the new stuff.
-   *
-   * @param model - The model to retrieve information for.
-   * @param max_tokens - The maximum number of tokens.
-   * @param system_fingerprint - The system fingerprint.
-   * @param useMaxTokens - Whether to use the maximum number of max tokens.
-   * @returns An object containing information about the new stuff.
-   *
-   * @author H0llyW00dzZ
-   */
-  private getNewStuff(
-    model: string,
-    max_tokens?: number,
-    system_fingerprint?: string,
-    useMaxTokens: boolean = true,
-  ): {
-    max_tokens?: number;
-    system_fingerprint?: string;
-    isNewModel: boolean;
-    payloadType: 'chat' | 'image';
-    isDalle: boolean;
-  } {
-    const modelConfig = {
-      ...useAppConfig.getState().modelConfig,
-      ...useChatStore.getState().currentSession().mask.modelConfig,
-    };
-    const isNewModel = model.endsWith("-preview");
-    const isDalle = model.startsWith("dall-e");
-    let payloadType: 'chat' | 'image' = 'chat';
-
-    if (isNewModel) {
-      return {
-        max_tokens: useMaxTokens ? (max_tokens !== undefined ? max_tokens : modelConfig.max_tokens) : undefined,
-        system_fingerprint:
-          system_fingerprint !== undefined
-            ? system_fingerprint
-            : modelConfig.system_fingerprint,
-        isNewModel: true,
-        payloadType,
-        isDalle,
-      };
-    } else if (isDalle) {
-      payloadType = 'image';
-    }
-
-    return {
-      isNewModel: false,
-      payloadType,
-      isDalle,
-    };
   }
 
   /**
@@ -225,8 +176,8 @@ export class ChatGPTApi implements LLMApi {
      * @usage in this chat: prompt
      * @example : A Best Picture of Andromeda Galaxy
      */
-    const actualModel = this.getModelForInstructVersion(modelConfig.model);
-    const { max_tokens, system_fingerprint } = this.getNewStuff(
+    const actualModel = getModelForInstructVersion(modelConfig.model);
+    const { max_tokens, system_fingerprint } = getNewStuff(
       modelConfig.model,
       modelConfig.max_tokens,
       modelConfig.system_fingerprint,
@@ -274,7 +225,7 @@ export class ChatGPTApi implements LLMApi {
      * 
      * @author H0llyW00dzZ
      */
-    const magicPayload = this.getNewStuff(defaultModel);
+    const magicPayload = getNewStuff(defaultModel);
     const provider = getProviderFromState();
 
     let payload;
@@ -667,21 +618,6 @@ export class ChatGPTApi implements LLMApi {
     }));
   }
 
-  /**
-   * Returns the model name for the given input model, accounting for instruct versions (For Instruct Version Still WIP).
-   * If the input model is not found in the model map, it returns the input model as is.
-   * @param inputModel - The input model name.
-   * @returns The corresponding model name or the input model if not found.
-   * @author H0llyW00dzZ
-   */
-
-  private getModelForInstructVersion(inputModel: string): string {
-    const modelMap: Record<string, string> = {
-      "dall-e-2-beta-instruct-vision": "dall-e-2",
-      "dall-e-3-beta-instruct-vision": "dall-e-3",
-    };
-    return modelMap[inputModel] || inputModel;
-  }
   /**
    * Saves the image from the response to the local filesystem.
    * @param imageResponse - The response containing the image data.
